+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `--stdin-filepath` flag to `buf format`, which reads a single `.proto` file from
   stdin and writes the formatted result to stdout. The path is not read from disk, and is
   only used to report parse errors and diffs.
+- Fix lint comment ignores on proto2 `group` fields being ignored.
 
 ## [v1.73.0] - 2026-09-11